# ecrpc/external_coordinator.pb.gw.go
// Code generated by protoc-gen-grpc-gateway. DO NOT EDIT.
// source: ecrpc/external_coordinator.proto

/*
Package ecrpc is a reverse proxy.

It translates gRPC into RESTful JSON APIs.
*/
package ecrpc

import (
	"context"
	"io"
	"net/http"

	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
	"github.com/grpc-ecosystem/grpc-gateway/v2/utilities"
	"google.golang.org/grpc"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/grpclog"
	"google.golang.org/grpc/metadata"
	"google.golang.org/grpc/status"
	"google.golang.org/protobuf/proto"
)

// Suppress "imported and not used" errors
var _ codes.Code
var _ io.Reader
var _ status.Status
var _ = runtime.String
var _ = utilities.NewDoubleArray
var _ = metadata.Join

func request_ExternalCoordinator_RegisterMissionControl_0(ctx context.Context, marshaler runtime.Marshaler, client ExternalCoordinatorClient, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq RegisterMissionControlRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := client.RegisterMissionControl(ctx, &protoReq, grpc.Header(&metadata.HeaderMD), grpc.Trailer(&metadata.TrailerMD))
	return msg, metadata, err

}

func local_request_ExternalCoordinator_RegisterMissionControl_0(ctx context.Context, marshaler runtime.Marshaler, server ExternalCoordinatorServer, req *http.Request, pathParams map[string]string) (proto.Message, runtime.ServerMetadata, error) {
	var protoReq RegisterMissionControlRequest
	var metadata runtime.ServerMetadata

	if err := marshaler.NewDecoder(req.Body).Decode(&protoReq); err != nil && err != io.EOF {
		return nil, metadata, status.Errorf(codes.InvalidArgument, "%v", err)
	}

	msg, err := server.RegisterMissionControl(ctx, &protoReq)
	return msg, metadata, err

}

func request_ExternalCoordinator_QueryAggregatedMissionControl_0(ctx context.Context, marshaler runtime.Marshaler, client ExternalCoordinatorClient, req *http.Request, pathParams map[string]string) (ExternalCoordinator_QueryAggregatedMissionControlClient, runtime.ServerMetadata, error) {
	var protoReq QueryAggregatedMissionControlRequest
	var metadata runtime.ServerMetadata

	stream, err := client.QueryAggregatedMissionControl(ctx, &protoReq)
	if err != nil {
		return nil, metadata, err
	}
	header, err := stream.Header()
	if err != nil {
		return nil, metadata, err
	}
	metadata.HeaderMD = header
	return stream, metadata, nil

}

// RegisterExternalCoordinatorHandlerServer registers the http handlers for service ExternalCoordinator to "mux".
// UnaryRPC     :call ExternalCoordinatorServer directly.
// StreamingRPC :currently unsupported pending https://github.com/grpc/grpc-go/issues/906.
// Note that using this registration option will cause many gRPC library features to stop working. Consider using RegisterExternalCoordinatorHandlerFromEndpoint instead.
func RegisterExternalCoordinatorHandlerServer(ctx context.Context, mux *runtime.ServeMux, server ExternalCoordinatorServer) error {

	mux.Handle("POST", pattern_ExternalCoordinator_RegisterMissionControl_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		var stream runtime.ServerTransportStream
		ctx = grpc.NewContextWithServerTransportStream(ctx, &stream)
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateIncomingContext(ctx, mux, req, "/ecrpc.ExternalCoordinator/RegisterMissionControl", runtime.WithHTTPPathPattern("/v1/register_mission_control"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := local_request_ExternalCoordinator_RegisterMissionControl_0(annotatedContext, inboundMarshaler, server, req, pathParams)
		md.HeaderMD, md.TrailerMD = metadata.Join(md.HeaderMD, stream.Header()), metadata.Join(md.TrailerMD, stream.Trailer())
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_ExternalCoordinator_RegisterMissionControl_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_ExternalCoordinator_QueryAggregatedMissionControl_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		err := status.Error(codes.Unimplemented, "streaming calls are not yet supported in the in-process transport")
		_, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
		return
	})

	return nil
}

// RegisterExternalCoordinatorHandlerFromEndpoint is same as RegisterExternalCoordinatorHandler but
// automatically dials to "endpoint" and closes the connection when "ctx" gets done.
func RegisterExternalCoordinatorHandlerFromEndpoint(ctx context.Context, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) (err error) {
	conn, err := grpc.NewClient(endpoint, opts...)
	if err != nil {
		return err
	}
	defer func() {
		if err != nil {
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
			return
		}
		go func() {
			<-ctx.Done()
			if cerr := conn.Close(); cerr != nil {
				grpclog.Errorf("Failed to close conn to %s: %v", endpoint, cerr)
			}
		}()
	}()

	return RegisterExternalCoordinatorHandler(ctx, mux, conn)
}

// RegisterExternalCoordinatorHandler registers the http handlers for service ExternalCoordinator to "mux".
// The handlers forward requests to the grpc endpoint over "conn".
func RegisterExternalCoordinatorHandler(ctx context.Context, mux *runtime.ServeMux, conn *grpc.ClientConn) error {
	return RegisterExternalCoordinatorHandlerClient(ctx, mux, NewExternalCoordinatorClient(conn))
}

// RegisterExternalCoordinatorHandlerClient registers the http handlers for service ExternalCoordinator
// to "mux". The handlers forward requests to the grpc endpoint over the given implementation of "ExternalCoordinatorClient".
// Note: the gRPC framework executes interceptors within the gRPC handler. If the passed in "ExternalCoordinatorClient"
// doesn't go through the normal gRPC flow (creating a gRPC client etc.) then it will be up to the passed in
// "ExternalCoordinatorClient" to call the correct interceptors.
func RegisterExternalCoordinatorHandlerClient(ctx context.Context, mux *runtime.ServeMux, client ExternalCoordinatorClient) error {

	mux.Handle("POST", pattern_ExternalCoordinator_RegisterMissionControl_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/ecrpc.ExternalCoordinator/RegisterMissionControl", runtime.WithHTTPPathPattern("/v1/register_mission_control"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_ExternalCoordinator_RegisterMissionControl_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_ExternalCoordinator_RegisterMissionControl_0(annotatedContext, mux, outboundMarshaler, w, req, resp, mux.GetForwardResponseOptions()...)

	})

	mux.Handle("GET", pattern_ExternalCoordinator_QueryAggregatedMissionControl_0, func(w http.ResponseWriter, req *http.Request, pathParams map[string]string) {
		ctx, cancel := context.WithCancel(req.Context())
		defer cancel()
		inboundMarshaler, outboundMarshaler := runtime.MarshalerForRequest(mux, req)
		var err error
		var annotatedContext context.Context
		annotatedContext, err = runtime.AnnotateContext(ctx, mux, req, "/ecrpc.ExternalCoordinator/QueryAggregatedMissionControl", runtime.WithHTTPPathPattern("/v1/query_aggregated_mission_control"))
		if err != nil {
			runtime.HTTPError(ctx, mux, outboundMarshaler, w, req, err)
			return
		}
		resp, md, err := request_ExternalCoordinator_QueryAggregatedMissionControl_0(annotatedContext, inboundMarshaler, client, req, pathParams)
		annotatedContext = runtime.NewServerMetadataContext(annotatedContext, md)
		if err != nil {
			runtime.HTTPError(annotatedContext, mux, outboundMarshaler, w, req, err)
			return
		}

		forward_ExternalCoordinator_QueryAggregatedMissionControl_0(annotatedContext, mux, outboundMarshaler, w, req, func() (proto.Message, error) { return resp.Recv() }, mux.GetForwardResponseOptions()...)

	})

	return nil
}

var (
	pattern_ExternalCoordinator_RegisterMissionControl_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "register_mission_control"}, ""))

	pattern_ExternalCoordinator_QueryAggregatedMissionControl_0 = runtime.MustPattern(runtime.NewPattern(1, []int{2, 0, 2, 1}, []string{"v1", "query_aggregated_mission_control"}, ""))
)

var (
	forward_ExternalCoordinator_RegisterMissionControl_0 = runtime.ForwardResponseMessage

	forward_ExternalCoordinator_QueryAggregatedMissionControl_0 = runtime.ForwardResponseStream
)